Introduce patchset-reference to improve reproducibility.
Rationale: I used to use a lot of local development to develop and use
my system. I wanted to improve reproducibility by being able to
properly pin patchsets that I use. This is only the first step of a
more systematic rewrite of currently used patchsets.

This commit has been produced in the following environment:
/home/graves/spheres/info/.bare/guix.git:182ef1677fb05292ed3f81cb96ac5ba73b10092a
/home/graves/spheres/info/.bare/nonguix.git:a664b6da35513fe007c1d48936e74541074e01fc
/home/graves/spheres/info/.bare/rde.git:47159132b22807cfb60152ba5d37755822bf67ff